Move bot protection handling to handlerwebapp module
For upcoming usage in TakeBranchResultInput
